♻️ App has been extracted from ...dynamic + index
